scripts/screenshots/requirements.txt
playwright
pyyaml
